requirements.txt
langchain>=0.1.0
langchain-core>=0.1.17
langchain-openai>=0.0.5
langgraph>=0.0.5
openai>=1.10.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.7.0
pydantic>=2.0.0
numpy>=1.24.0
exa-py>=1.0.4
mem0ai>=1.0.0
